Commit 3: Part 2, lecturer-submit: Added functionality to allow decimal hours and set en-ZA for R currency; update views and controller with more efficient code to perfom their intended functionalities using the help of A.I

+++ Models/Claim.cs
@@ -12,7 +12,7 @@ namespace CMCSApp.Models
         public string LecturerName { get; set; }
 
         [Required]
-        public int HoursWorked { get; set; }
+        public decimal HoursWorked { get; set; }
 
         [Required, DataType(DataType.Currency)]
         public decimal HourlyRate { get; set; }
